refactor: extract calendar drawers
