+++ app/src/main/java/com/kgs/calendar/ui/KgsCalendarApp.kt
@@ -2214,7 +2214,7 @@ private data class LogoParticleSpec(
 )
 
 @Composable
-private fun KgsLogoBurstButton() {
+internal fun KgsLogoBurstButton() {
     val scope = rememberCoroutineScope()
     val burstProgress = remember { Animatable(1f) }
     var burstJob by remember { mutableStateOf<Job?>(null) }
@@ -2300,595 +2300,6 @@ private fun KgsLogoBurstButton() {
                 },
             contentScale = ContentScale.Fit,
         )
-    }
-}
-
-@Composable
-private fun CalendarDrawer(
-    visible: Boolean,
-    state: CalendarUiState,
-    onDismiss: () -> Unit,
-    onViewSelected: (CalendarViewMode) -> Unit,
-    onSync: () -> Unit,
-    onCollectionVisibleInViews: (String, Boolean) -> Unit,
-    onCollectionSettings: (CollectionEntity) -> Unit,
-    onAppSettings: () -> Unit,
-    problems: List<ProblemItem>,
-    onProblems: () -> Unit,
-) {
-    val drawerProgress by animateFloatAsState(
-        targetValue = if (visible) 1f else 0f,
-        animationSpec = tween(if (visible) 135 else 115, easing = if (visible) MotionStandard else MotionStandardAccelerate),
-        label = "drawerProgress",
-    )
-    val quietInteraction = remember { MutableInteractionSource() }
-    val density = LocalDensity.current
-    val drawerWidth = 292.dp
-    val drawerWidthPx = with(density) { drawerWidth.toPx() }
-    var dragOffsetPx by remember { mutableFloatStateOf(0f) }
-    val effectiveProgress = (drawerProgress + dragOffsetPx / drawerWidthPx).coerceIn(0f, 1f)
-    LaunchedEffect(visible) {
-        if (visible) {
-            dragOffsetPx = 0f
-        } else {
-            delay(150)
-            dragOffsetPx = 0f
-        }
-    }
-    val statusTop = WindowInsets.statusBars.asPaddingValues().calculateTopPadding()
-    val navBottom = WindowInsets.navigationBars.asPaddingValues().calculateBottomPadding()
-    Box(Modifier.fillMaxSize()) {
-        if (effectiveProgress > 0.01f) {
-            Box(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .alpha(effectiveProgress)
-                    .background(Color.Black.copy(alpha = 0.58f))
-                    .clickable(
-                        interactionSource = quietInteraction,
-                        indication = null,
-                        onClick = onDismiss,
-                    ),
-            )
-        }
-        Surface(
-            modifier = Modifier
-                .width(drawerWidth)
-                .fillMaxHeight()
-                .offset {
-                    IntOffset(((-drawerWidthPx) * (1f - effectiveProgress)).roundToInt(), 0)
-                }
-                .pointerInput(visible, drawerWidthPx) {
-                    detectHorizontalDragGestures(
-                        onDragCancel = { dragOffsetPx = 0f },
-                        onDragEnd = {
-                            if (dragOffsetPx <= -drawerWidthPx * 0.28f) {
-                                onDismiss()
-                            } else {
-                                dragOffsetPx = 0f
-                            }
-                        },
-                    ) { change, dragAmount ->
-                        if (dragAmount < 0f || dragOffsetPx < 0f) {
-                            change.consume()
-                            dragOffsetPx = (dragOffsetPx + dragAmount).coerceIn(-drawerWidthPx, 0f)
-                        }
-                    }
-                },
-            shape = RoundedCornerShape(topEnd = 18.dp, bottomEnd = 18.dp),
-            color = MaterialTheme.colorScheme.background,
-            shadowElevation = 0.dp,
-        ) {
-            Column(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .verticalScroll(rememberScrollState())
-                    .padding(top = statusTop + 22.dp, bottom = navBottom + 14.dp),
-            ) {
-                Row(
-                    modifier = Modifier.padding(horizontal = 22.dp, vertical = 12.dp),
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(12.dp),
-                ) {
-                    KgsLogoBurstButton()
-                    Text(
-                        "KGS Calendar",
-                        fontSize = 22.sp,
-                        lineHeight = 25.sp,
-                        fontWeight = FontWeight.ExtraBold,
-                        fontFamily = FontFamily.SansSerif,
-                        color = WarmInk,
-                    )
-                }
-                DrawerViewItem(stringResource(R.string.agenda), Icons.Default.ViewAgenda, CalendarViewMode.Agenda, state.selectedView, onViewSelected)
-                DrawerViewItem(stringResource(R.string.day), Icons.Default.ViewDay, CalendarViewMode.Day, state.selectedView, onViewSelected)
-                DrawerViewItem(stringResource(R.string.three_days), Icons.Default.ViewWeek, CalendarViewMode.ThreeDay, state.selectedView, onViewSelected)
-                DrawerViewItem(stringResource(R.string.month), Icons.Default.CalendarMonth, CalendarViewMode.Month, state.selectedView, onViewSelected)
-                HorizontalDivider(color = WarmLine, modifier = Modifier.padding(top = 10.dp))
-                DrawerActionRow(stringResource(R.string.refresh), Icons.Default.Refresh, onSync)
-                HorizontalDivider(color = WarmLine)
-                val accounts = state.accounts.ifEmpty { state.account?.let(::listOf).orEmpty() }
-                val drawerCollections = state.collections.filter { it.isEnabled }
-                val collectionsByAccount = drawerCollections.groupBy { it.accountId }
-                val visibleAccounts = accounts.filter { account ->
-                    collectionsByAccount[account.id].orEmpty().isNotEmpty()
-                }
-                visibleAccounts.forEachIndexed { index, account ->
-                    if (index > 0) {
-                        HorizontalDivider(color = WarmLine.copy(alpha = 0.62f), modifier = Modifier.padding(horizontal = 24.dp, vertical = 10.dp))
-                    }
-                    DrawerSourceSection(
-                        account = account,
-                        collections = collectionsByAccount[account.id].orEmpty(),
-                        hiddenCollectionHrefs = state.hiddenCollectionHrefs,
-                        onCollectionVisibleInViews = onCollectionVisibleInViews,
-                        onCollectionSettings = onCollectionSettings,
-                    )
-                }
-                HorizontalDivider(color = WarmLine, modifier = Modifier.padding(top = 12.dp))
-                DrawerActionRow(stringResource(R.string.app_settings), Icons.Default.Settings, onAppSettings)
-                if (problems.isNotEmpty()) {
-                    DrawerProblemRow(
-                        count = problems.size,
-                        onClick = onProblems,
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun DrawerProblemRow(count: Int, onClick: () -> Unit) {
-    val color = SyncPendingOrange
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 12.dp, vertical = 3.dp)
-            .clip(RoundedCornerShape(28.dp))
-            .background(color.copy(alpha = if (MaterialTheme.colorScheme.background.isDark()) 0.34f else 0.18f))
-            .clickable(onClick = onClick)
-            .padding(horizontal = 20.dp, vertical = 11.dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(22.dp),
-    ) {
-        Icon(
-            imageVector = Icons.Default.WarningAmber,
-            contentDescription = null,
-            tint = color,
-            modifier = Modifier.size(23.dp),
-        )
-        Column(Modifier.weight(1f), verticalArrangement = Arrangement.spacedBy(1.dp)) {
-            Text(stringResource(R.string.problems), color = WarmInk, fontSize = 15.sp, lineHeight = 18.sp, fontWeight = FontWeight.SemiBold)
-            Text(stringResource(R.string.problem_count, count), color = MaterialTheme.colorScheme.onSurfaceVariant, fontSize = 11.sp, lineHeight = 13.sp)
-        }
-    }
-}
-
-@Composable
-private fun DrawerSourceSection(
-    account: com.kgs.calendar.data.local.entity.AccountEntity,
-    collections: List<CollectionEntity>,
-    hiddenCollectionHrefs: Set<String>,
-    onCollectionVisibleInViews: (String, Boolean) -> Unit,
-    onCollectionSettings: (CollectionEntity) -> Unit,
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 24.dp, vertical = 12.dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(14.dp),
-    ) {
-        val initial = (account.displayName ?: account.username).take(1).ifBlank { "?" }
-        Box(Modifier.size(30.dp).clip(CircleShape).background(WarmBrown), contentAlignment = Alignment.Center) {
-            Text(initial, color = Color.White, fontSize = 13.sp, lineHeight = 15.sp, fontWeight = FontWeight.Bold)
-        }
-        Column(Modifier.weight(1f)) {
-            Text(account.displayName ?: account.username, color = WarmInk, fontSize = 13.sp, lineHeight = 16.sp, fontWeight = FontWeight.SemiBold, maxLines = 1, overflow = TextOverflow.Ellipsis)
-            val subtitle = when {
-                account.id == UiLocalAccountId || account.serverUrl.startsWith(UiLocalCollectionPrefix) -> stringResource(R.string.local_calendar)
-                account.isAndroidProviderForUi() -> stringResource(R.string.android_device_calendars)
-                account.sourceType == SourceType.ReadOnlyUrl || account.username == "Read-only URL" -> stringResource(R.string.read_only_url)
-                else -> stringResource(R.string.nextcloud)
-            }
-            Text(
-                subtitle,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                fontSize = 11.sp,
-                lineHeight = 13.sp,
-            )
-        }
-    }
-    val sortedCollections = collections.sortedWith(
-        compareBy<CollectionEntity> {
-            when {
-                it.supportsEvents && it.supportsTasks -> 0
-                it.supportsEvents -> 1
-                else -> 2
-            }
-        }.thenBy { it.displayName.lowercase(Locale.ROOT) },
-    )
-    sortedCollections.forEach { collection ->
-        DrawerCollectionRow(
-            collection = collection,
-            visibleInViews = collection.href !in hiddenCollectionHrefs,
-            onVisibleInViewsChanged = { onCollectionVisibleInViews(collection.href, it) },
-            onSettings = { onCollectionSettings(collection) },
-        )
-    }
-}
-
-@Composable
-private fun DrawerViewItem(
-    label: String,
-    icon: ImageVector,
-    viewMode: CalendarViewMode,
-    selectedView: CalendarViewMode,
-    onViewSelected: (CalendarViewMode) -> Unit,
-) {
-    val selected = viewMode == selectedView
-    val background by animateColorAsState(
-        targetValue = if (selected) WarmPeach else Color.Transparent,
-        animationSpec = tween(MotionMedium, easing = MotionStandard),
-        label = "drawerSelection",
-    )
-    val glyphScale by animateFloatAsState(
-        targetValue = if (selected) 1f else 0.92f,
-        animationSpec = tween(MotionMedium, easing = MotionStandard),
-        label = "drawerGlyphScale",
-    )
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 12.dp, vertical = 3.dp)
-            .clip(RoundedCornerShape(28.dp))
-            .background(background)
-            .clickable { onViewSelected(viewMode) }
-            .padding(horizontal = 20.dp, vertical = 11.dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(22.dp),
-    ) {
-        Icon(
-            imageVector = icon,
-            contentDescription = null,
-            tint = if (selected) WarmInk else MaterialTheme.colorScheme.onSurfaceVariant,
-            modifier = Modifier.size(23.dp).scale(glyphScale),
-        )
-        Text(label, color = WarmInk, fontSize = 15.sp, lineHeight = 18.sp, fontWeight = if (selected) FontWeight.SemiBold else FontWeight.Normal)
-    }
-}
-
-@Composable
-private fun DrawerActionRow(label: String, icon: ImageVector, onClick: () -> Unit) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable(onClick = onClick)
-            .padding(horizontal = 30.dp, vertical = 14.dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(24.dp),
-    ) {
-        Icon(icon, contentDescription = null, tint = WarmInk, modifier = Modifier.size(23.dp))
-        Text(label, color = WarmInk, fontSize = 14.sp, lineHeight = 17.sp)
-    }
-}
-
-@Composable
-private fun DrawerCollectionGroupHeader(label: String) {
-    Text(
-        label,
-        color = MaterialTheme.colorScheme.onSurfaceVariant,
-        fontSize = 11.sp,
-        lineHeight = 13.sp,
-        fontWeight = FontWeight.SemiBold,
-        modifier = Modifier.padding(start = 30.dp, end = 24.dp, top = 8.dp, bottom = 4.dp),
-    )
-}
-
-@Composable
-private fun DrawerCollectionRow(
-    collection: CollectionEntity,
-    visibleInViews: Boolean,
-    onVisibleInViewsChanged: (Boolean) -> Unit,
-    onSettings: () -> Unit,
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable(onClick = onSettings)
-            .padding(horizontal = 24.dp, vertical = 9.dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(18.dp),
-    ) {
-        CollectionToggle(color = Color(collection.color), checked = visibleInViews) {
-            onVisibleInViewsChanged(!visibleInViews)
-        }
-        Column(Modifier.weight(1f)) {
-            Text(collection.displayName, color = WarmInk, fontSize = 14.sp, lineHeight = 17.sp, maxLines = 1, overflow = TextOverflow.Ellipsis)
-            Row(
-                modifier = Modifier.padding(top = 3.dp),
-                horizontalArrangement = Arrangement.spacedBy(5.dp),
-                verticalAlignment = Alignment.CenterVertically,
-            ) {
-                if (collection.supportsEvents) {
-                    CalendarCapabilityChip(stringResource(R.string.events), Color(0xFF2F5AEA))
-                }
-                if (collection.supportsTasks) {
-                    CalendarCapabilityChip(stringResource(R.string.tasks), Color(0xFF00A86B))
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun HiddenCalendarCreationDialog(
-    collection: CollectionEntity,
-    itemLabel: String,
-    onDismiss: () -> Unit,
-    onUnhide: () -> Unit,
-) {
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = {
-            Text(stringResource(R.string.hidden_calendar_creation_title))
-        },
-        text = {
-            Column(verticalArrangement = Arrangement.spacedBy(14.dp)) {
-                Text(
-                    stringResource(R.string.hidden_calendar_creation_body, itemLabel, collection.displayName),
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    fontSize = 14.sp,
-                    lineHeight = 18.sp,
-                )
-                val hiddenCalendarShape = RoundedCornerShape(18.dp)
-                Surface(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .dashedBorder(SyncPendingOrange, 18.dp)
-                        .clickable(onClick = onUnhide),
-                    shape = hiddenCalendarShape,
-                    color = MaterialTheme.colorScheme.surface,
-                ) {
-                    Row(
-                        modifier = Modifier.padding(horizontal = 14.dp, vertical = 12.dp),
-                        verticalAlignment = Alignment.CenterVertically,
-                        horizontalArrangement = Arrangement.spacedBy(14.dp),
-                    ) {
-                        CollectionToggle(color = Color(collection.color), checked = false, onClick = onUnhide)
-                        Column(Modifier.weight(1f)) {
-                            Text(
-                                collection.displayName,
-                                color = WarmInk,
-                                fontSize = 14.sp,
-                                lineHeight = 17.sp,
-                                fontWeight = FontWeight.SemiBold,
-                                maxLines = 1,
-                                overflow = TextOverflow.Ellipsis,
-                            )
-                            Row(
-                                modifier = Modifier.padding(top = 4.dp),
-                                horizontalArrangement = Arrangement.spacedBy(5.dp),
-                                verticalAlignment = Alignment.CenterVertically,
-                            ) {
-                                if (collection.supportsEvents) {
-                                    CalendarCapabilityChip(stringResource(R.string.events), Color(0xFF2F5AEA))
-                                }
-                                if (collection.supportsTasks) {
-                                    CalendarCapabilityChip(stringResource(R.string.tasks), Color(0xFF00A86B))
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        },
-        confirmButton = {
-            TextButton(onClick = onUnhide) {
-                Text(stringResource(R.string.unhide_calendar))
-            }
-        },
-        dismissButton = {
-            TextButton(onClick = onDismiss) {
-                Text(stringResource(R.string.close))
-            }
-        },
-    )
-}
-
-@Composable
-private fun CalendarCapabilityChip(label: String, color: Color) {
-    Surface(
-        shape = RoundedCornerShape(999.dp),
-        color = color.copy(alpha = if (MaterialTheme.colorScheme.background.isDark()) 0.30f else 0.14f),
-        border = BorderStroke(1.dp, color.copy(alpha = 0.42f)),
-    ) {
-        Text(
-            label,
-            color = color,
-            fontSize = 10.sp,
-            lineHeight = 12.sp,
-            fontWeight = FontWeight.SemiBold,
-            modifier = Modifier.padding(horizontal = 7.dp, vertical = 2.dp),
-        )
-    }
-}
-
-@Composable
-private fun CollectionToggle(color: Color, checked: Boolean, onClick: () -> Unit) {
-    val fillColor by animateColorAsState(
-        targetValue = if (checked) color else MaterialTheme.colorScheme.surface.copy(alpha = 0.72f),
-        animationSpec = tween(MotionMedium, easing = MotionStandard),
-        label = "collectionToggleFill",
-    )
-    val borderColor by animateColorAsState(
-        targetValue = if (checked) color else MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.28f),
-        animationSpec = tween(MotionMedium, easing = MotionStandard),
-        label = "collectionToggleBorder",
-    )
-    val checkScale by animateFloatAsState(
-        targetValue = if (checked) 1f else 0f,
-        animationSpec = tween(MotionShort, easing = MotionStandard),
-        label = "collectionToggleCheck",
-    )
-    val shape = RoundedCornerShape(7.dp)
-    Box(
-        modifier = Modifier
-            .size(23.dp)
-            .clip(shape)
-            .background(fillColor)
-            .border(1.2.dp, borderColor, shape)
-            .clickable(onClick = onClick),
-        contentAlignment = Alignment.Center,
-    ) {
-        Icon(
-            Icons.Default.Check,
-            contentDescription = null,
-            tint = Color.White,
-            modifier = Modifier
-                .size(16.dp)
-                .scale(checkScale),
-        )
-    }
-}
-
-@Composable
-private fun DrawerGlyph(selected: Boolean, modifier: Modifier = Modifier) {
-    val color = if (selected) WarmInk else MaterialTheme.colorScheme.onSurfaceVariant
-    Column(
-        modifier = modifier
-            .width(24.dp)
-            .height(22.dp),
-        verticalArrangement = Arrangement.spacedBy(2.dp),
-    ) {
-        repeat(3) {
-            Box(
-                Modifier
-                    .fillMaxWidth()
-                    .weight(1f)
-                    .clip(RoundedCornerShape(2.dp))
-                    .background(color.copy(alpha = if (selected) 1f else 0.72f)),
-            )
-        }
-    }
-}
-
-@Composable
-private fun TaskDrawer(
-    visible: Boolean,
-    state: CalendarUiState,
-    onDismiss: () -> Unit,
-    onTaskStatusChanged: (String, String) -> Unit,
-    onTaskClick: (TaskEntity) -> Unit,
-    onShowCompleted: () -> Unit,
-    onCreateTask: () -> Unit,
-) {
-    val drawerProgress by animateFloatAsState(
-        targetValue = if (visible) 1f else 0f,
-        animationSpec = tween(if (visible) 145 else 115, easing = if (visible) MotionStandard else MotionStandardAccelerate),
-        label = "taskDrawerProgress",
-    )
-    val quietInteraction = remember { MutableInteractionSource() }
-    val density = LocalDensity.current
-    val drawerWidth = 336.dp
-    val drawerWidthPx = with(density) { drawerWidth.toPx() }
-    var dragOffsetPx by remember { mutableFloatStateOf(0f) }
-    val effectiveProgress = (drawerProgress - dragOffsetPx / drawerWidthPx).coerceIn(0f, 1f)
-    LaunchedEffect(visible) {
-        if (visible) {
-            dragOffsetPx = 0f
-        } else {
-            delay(150)
-            dragOffsetPx = 0f
-        }
-    }
-    val statusTop = WindowInsets.statusBars.asPaddingValues().calculateTopPadding()
-    val navBottom = WindowInsets.navigationBars.asPaddingValues().calculateBottomPadding()
-    Box(Modifier.fillMaxSize()) {
-        if (effectiveProgress > 0.01f) {
-            Box(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .alpha(effectiveProgress)
-                    .background(Color.Black.copy(alpha = 0.42f))
-                    .clickable(
-                        interactionSource = quietInteraction,
-                        indication = null,
-                        onClick = onDismiss,
-                    ),
-            )
-        }
-        Surface(
-            modifier = Modifier
-                .fillMaxHeight()
-                .width(drawerWidth)
-                .align(Alignment.CenterEnd)
-                .offset {
-                    IntOffset((drawerWidthPx * (1f - effectiveProgress)).roundToInt(), 0)
-                }
-                .pointerInput(visible, drawerWidthPx) {
-                    detectHorizontalDragGestures(
-                        onDragCancel = { dragOffsetPx = 0f },
-                        onDragEnd = {
-                            if (dragOffsetPx >= drawerWidthPx * 0.28f) {
-                                onDismiss()
-                            } else {
-                                dragOffsetPx = 0f
-                            }
-                        },
-                    ) { change, dragAmount ->
-                        if (dragAmount > 0f || dragOffsetPx > 0f) {
-                            change.consume()
-                            dragOffsetPx = (dragOffsetPx + dragAmount).coerceIn(0f, drawerWidthPx)
-                        }
-                    }
-                },
-            shape = RoundedCornerShape(topStart = 18.dp, bottomStart = 18.dp),
-            color = MaterialTheme.colorScheme.background,
-            shadowElevation = 0.dp,
-        ) {
-            Column(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(top = statusTop + 18.dp),
-            ) {
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(horizontal = 28.dp, vertical = 10.dp),
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(12.dp),
-                ) {
-                    Icon(Icons.Default.CheckCircle, contentDescription = null, tint = WarmInk, modifier = Modifier.size(24.dp))
-                    Column(Modifier.weight(1f)) {
-                        Text(stringResource(R.string.tasks), color = WarmInk, fontSize = 21.sp, lineHeight = 25.sp, fontWeight = FontWeight.SemiBold)
-                    }
-                    IconButton(
-                        onClick = onCreateTask,
-                        modifier = Modifier
-                            .size(38.dp)
-                            .clip(CircleShape)
-                            .background(accentContainerColor()),
-                    ) {
-                        Icon(
-                            Icons.Default.Add,
-                            contentDescription = stringResource(R.string.new_task),
-                            tint = accentContainerContentColor(),
-                            modifier = Modifier.size(22.dp),
-                        )
-                    }
-                }
-                TaskInbox(
-                    state = state,
-                    onTaskStatusChanged = onTaskStatusChanged,
-                    onDetail = { detail ->
-                        if (detail is DetailSheet.Task) onTaskClick(detail.task)
-                    },
-                    onShowCompleted = onShowCompleted,
-                )
-            }
-        }
     }
 }
 
@@ -10359,14 +9770,14 @@ internal fun CalendarUiState.allKnownCategoryTags(): List<String> =
         completedTasks.forEach { addAll(it.categories.toCategoryTags()) }
     }.distinctBy { it.lowercase(Locale.ROOT) }.sortedBy { it.lowercase(Locale.ROOT) }
 
-private data class ProblemItem(
+internal data class ProblemItem(
     val id: String,
     val title: String,
     val body: String,
     val target: ProblemTarget? = null,
 )
 
-private sealed interface ProblemTarget {
+internal sealed interface ProblemTarget {
     data class Event(val event: EventEntity) : ProblemTarget
     data class Task(val task: TaskEntity) : ProblemTarget
 }
@@ -11145,7 +10556,7 @@ private fun String.isLocalCollectionHrefUi(): Boolean = startsWith(UiLocalCollec
 private fun CollectionEntity.isAndroidProviderForUi(): Boolean =
     sourceType == SourceType.AndroidProvider || href.startsWith(UiAndroidCollectionPrefix)
 
-private fun AccountEntity.isAndroidProviderForUi(): Boolean =
+internal fun AccountEntity.isAndroidProviderForUi(): Boolean =
     sourceType == SourceType.AndroidProvider || id == UiAndroidAccountId
 
 internal data class EventEditorCapabilities(
